+++ app/index.tsx
@@ -79,6 +79,17 @@ export default function Index({ children }: RootLayoutProps) {
     [address, isConnected]
   )
 
+  const connectBitcoin = async () => {
+    const w = window as any
+    console.log("connect bitcoin")
+    if ("xfi" in w && w.xfi?.bitcoin) {
+      w.xfi.bitcoin.changeNetwork("testnet")
+      const btc = (await w.xfi.bitcoin.getAccounts())[0]
+      await setBitcoinAddress(btc)
+      // fetchBalances(true, btc)
+    }
+  }
+
   const fetchStakingDelegations = useCallback(
     debounce(async () => {
       try {
@@ -367,6 +378,7 @@ export default function Index({ children }: RootLayoutProps) {
           balancesLoading,
           balancesRefreshing,
           fees,
+          connectBitcoin,
           pools,
           poolsLoading,
           validators,
